Move WriteOutcome<T> from Trellis.Asp to Trellis.Results (Phase 1a PR6)
Per v2-redesign-plan.md §3.1, Application-layer repositories declare
Result<WriteOutcome<T>> return types. Application cannot reference
Trellis.Asp under Clean Architecture, so WriteOutcome<T> moves to
Trellis.Results (its dependencies RepresentationMetadata,
RetryAfterValue, EntityTagValue already live there).

Changes:
- Move Trellis.Asp/src/WriteOutcome.cs to Trellis.Results/src/.
- Change namespace Trellis.Asp -> Trellis.
- Soften doc comments to be transport-neutral (cases align with
  RFC 9110 outcomes but the type takes no transport dependency).
- Add per-case <param> XML docs for Created/Updated/Accepted etc.
- Move WriteOutcome<T> reference docs from trellis-api-asp.md to
  trellis-api-results.md; cross-link from the ASP doc.
- Add Phase 1a 'Breaking changes from v1' table entry for the
  package + namespace move.

WriteOutcomeExtensions, ToActionResult, ToHttpResult, and the
Prefer-aware Update helpers stay in Trellis.Asp - they translate
the transport-agnostic outcome to ASP.NET-specific responses.

Build: 0 warnings, 0 errors. Tests: 4928 passed, 15 skipped, 0 failed.
