+++ docs/api_reference/trellis-api-asp.md
@@ -396,107 +396,7 @@ public static class ValidationErrorsContext
 | `public static IDisposable BeginScope()` | `IDisposable` | Starts a new async-local validation collection scope; disposing restores the previous scope and property name. |
 | `public static ValidationError? GetValidationError()` | `ValidationError?` | Returns the aggregated `ValidationError` for the current scope, or `null` when no errors were collected. |
 
-### `WriteOutcome<T>`
-
-**Declaration**
-
-```csharp
-public abstract record WriteOutcome<T>
-```
-
-| Name | Type | Description |
-| --- | --- | --- |
-| — | — | Base union type for write-operation responses. |
-
-| Signature | Returns | Description |
-| --- | --- | --- |
-| — | — | No public methods. |
-
-### `WriteOutcome<T>.Created`
-
-**Declaration**
-
-```csharp
-public sealed record Created(T Value, string Location, RepresentationMetadata? Metadata = null) : WriteOutcome<T>;
-```
-
-| Name | Type | Description |
-| --- | --- | --- |
-| `Value` | `T` | Created resource or response body. |
-| `Location` | `string` | Absolute or relative location URI for the created resource. |
-| `Metadata` | `RepresentationMetadata?` | Optional representation metadata applied to the response. |
-
-| Signature | Returns | Description |
-| --- | --- | --- |
-| — | — | No public methods beyond record-generated members. |
-
-### `WriteOutcome<T>.Updated`
-
-**Declaration**
-
-```csharp
-public sealed record Updated(T Value, RepresentationMetadata? Metadata = null) : WriteOutcome<T>;
-```
-
-| Name | Type | Description |
-| --- | --- | --- |
-| `Value` | `T` | Updated resource or response body. |
-| `Metadata` | `RepresentationMetadata?` | Optional representation metadata applied to the response. |
-
-| Signature | Returns | Description |
-| --- | --- | --- |
-| — | — | No public methods beyond record-generated members. |
-
-### `WriteOutcome<T>.UpdatedNoContent`
-
-**Declaration**
-
-```csharp
-public sealed record UpdatedNoContent(RepresentationMetadata? Metadata = null) : WriteOutcome<T>;
-```
-
-| Name | Type | Description |
-| --- | --- | --- |
-| `Metadata` | `RepresentationMetadata?` | Optional representation metadata applied to the `204` response. |
-
-| Signature | Returns | Description |
-| --- | --- | --- |
-| — | — | No public methods beyond record-generated members. |
-
-### `WriteOutcome<T>.Accepted`
-
-**Declaration**
-
-```csharp
-public sealed record Accepted(T StatusBody, string? MonitorUri = null, RetryAfterValue? RetryAfter = null) : WriteOutcome<T>;
-```
-
-| Name | Type | Description |
-| --- | --- | --- |
-| `StatusBody` | `T` | Accepted-operation status payload. |
-| `MonitorUri` | `string?` | Optional monitor URI written to `Location` / `Accepted(...)`. |
-| `RetryAfter` | `RetryAfterValue?` | Optional `Retry-After` header value. |
-
-| Signature | Returns | Description |
-| --- | --- | --- |
-| — | — | No public methods beyond record-generated members. |
-
-### `WriteOutcome<T>.AcceptedNoContent`
-
-**Declaration**
-
-```csharp
-public sealed record AcceptedNoContent(string? MonitorUri = null, RetryAfterValue? RetryAfter = null) : WriteOutcome<T>;
-```
-
-| Name | Type | Description |
-| --- | --- | --- |
-| `MonitorUri` | `string?` | Optional monitor URI written to `Location` / `Accepted(...)`. |
-| `RetryAfter` | `RetryAfterValue?` | Optional `Retry-After` header value. |
-
-| Signature | Returns | Description |
-| --- | --- | --- |
-| — | — | No public methods beyond record-generated members. |
+> **Note:** `WriteOutcome<T>` and its case records (`Created`, `Updated`, `UpdatedNoContent`, `Accepted`, `AcceptedNoContent`) live in `Trellis.Results` (namespace `Trellis`) — see [`trellis-api-results.md`](./trellis-api-results.md). They are documented there because Application-layer repositories (which cannot reference `Trellis.Asp`) must declare them as return types. The `WriteOutcomeExtensions` mappers below are the ASP.NET-specific surface that translates each case to `ActionResult` / `IResult`.
 
 ### `WriteOutcomeExtensions`
